+++ client/src/containers/DashboardContainer.jsx
@@ -1,10 +1,11 @@
 import React from 'react';
+import {Link} from 'react-router';
 import adminLte from 'adminlte-reactjs';
-const StatTile = adminLte.StatTile;
+const InfoTile = adminLte.InfoTile;
 
 let DashboardContainer = React.createClass({
   getInitialState() {
-    return {fontsCount: 0};
+    return {fontsCount: '0'};
   },
   componentDidMount() {
     fetch('/api/fonts/count')
@@ -22,14 +23,16 @@ let DashboardContainer = React.createClass({
           <h1>Dashboard</h1>
         </section>
         <section className="content">
-          <StatTile
-            width='2'
-            theme='bg-yellow'
-            icon='fa-font'
-            subject='Fonts'
-            stats={String(this.state.count)}
-            link='/fonts'
-          />
+          <Link to="/fonts">
+            <InfoTile
+              width='2'
+              content=''
+              icon='fa-font'
+              stats={String(this.state.count)}
+              subject='Fonts'
+              theme='bg-aqua'
+            />
+          </Link>
         </section>
       </div>
     );